requirements.txt
# This file was autogenerated by uv via the following command:
#    uv export --no-hashes --no-editable --format requirements-txt
.
annotated-types==0.7.0
    # via pydantic
blinker==1.9.0
    # via flask
click==8.3.2
    # via flask
colorama==0.4.6 ; sys_platform == 'win32'
    # via click
et-xmlfile==2.0.0
    # via openpyxl
flask==3.1.3
    # via
    #   flask-cors
    #   servicoscrc
flask-cors==6.0.2
    # via servicoscrc
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   werkzeug
openpyxl==3.1.5
    # via servicoscrc
pydantic==2.12.5
    # via servicoscrc
pydantic-core==2.41.5
    # via pydantic
python-dotenv==1.2.2
    # via servicoscrc
typing-extensions==4.15.0
    # via
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
werkzeug==3.1.8
    # via
    #   flask
    #   flask-cors
